VoiceBot/requirements.txt
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
comtypes==1.1.11
frozenlist==1.3.0
idna==3.3
langdetect==1.0.9
multidict==6.0.2
pip==21.3.1
pydub==0.25.1
pypiwin32==223
python-dotenv==0.20.0
pyttsx3==2.90
pytz==2022.1
pywin32==304
setuptools==60.2.0
six==1.16.0
wheel==0.37.1
yarl==1.7.2
